// prisma/generated/zod/inputTypeSchemas/BookingRequestWhereInputSchema.ts
export const BookingRequestWhereInputSchema: z.ZodType<Prisma.BookingRequestWhereInput> = z.object({
  AND: z.union([ z.lazy(() => BookingRequestWhereInputSchema),z.lazy(() => BookingRequestWhereInputSchema).array() ]).optional(),
  OR: z.lazy(() => BookingRequestWhereInputSchema).array().optional(),
  NOT: z.union([ z.lazy(() => BookingRequestWhereInputSchema),z.lazy(() => BookingRequestWhereInputSchema).array() ]).optional(),
  id: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  merchantId: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  prefilledFields: z.lazy(() => JsonNullableFilterSchema).optional(),
  customerId: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  resourceId: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  templateId: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  expiresAt: z.union([ z.lazy(() => DateTimeNullableFilterSchema),z.coerce.date() ]).optional().nullable(),
  isActive: z.union([ z.lazy(() => BoolFilterSchema),z.boolean() ]).optional(),
  notes: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  status: z.union([ z.lazy(() => EnumBookingRequestStatusFilterSchema),z.lazy(() => BookingRequestStatusSchema) ]).optional(),
  actionStatus: z.lazy(() => JsonFilterSchema).optional(),
  finalizedSnapshot: z.lazy(() => JsonNullableFilterSchema).optional(),
  createdAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  updatedAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  merchant: z.union([ z.lazy(() => MerchantScalarRelationFilterSchema),z.lazy(() => MerchantWhereInputSchema) ]).optional(),
  customer: z.union([ z.lazy(() => CustomerNullableScalarRelationFilterSchema),z.lazy(() => CustomerWhereInputSchema) ]).optional().nullable(),
  resource: z.union([ z.lazy(() => ResourceNullableScalarRelationFilterSchema),z.lazy(() => ResourceWhereInputSchema) ]).optional().nullable(),
  template: z.union([ z.lazy(() => BookingTemplateScalarRelationFilterSchema),z.lazy(() => BookingTemplateWhereInputSchema) ]).optional(),
  BookingRequestFlow: z.lazy(() => BookingRequestFlowListRelationFilterSchema).optional(),
  BookingRequestActionValue: z.lazy(() => BookingRequestActionValueListRelationFilterSchema).optional(),
  Booking: z.lazy(() => BookingListRelationFilterSchema).optional()
}).strict();